+++ pdf/lib/src/widgets/content.dart
@@ -151,7 +151,7 @@ class TableOfContent extends StatelessWidget {
                   thickness: 0.2,
                 )),
                 SizedBox(width: 8),
-                Text('${c.page}'),
+                DelayedWidget(build: (_) => Text('${c.page}')),
               ],
             ),
           ),